refactored User class name to UserModel
